Add a header-includes option
By allowing users to include arbitrary content in the header-includes
field, they can include whatever additional content they might want in
the poster's header. I made this change so that I could use a
sans-serif font for the whole poster. This commit includes a description
and example usage of header-includes in the README.

This may be relevant for
https://github.com/brentthorne/posterdown/issues/45